% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{adjperf}
\alias{adjperf}
\title{Extract adjusted energy cost}
\usage{
adjperf(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
measurements adjusted for background in the form of \code{\link{retperf}}
}
\description{
The energy values are adjusted such that
background is shifted to the `base` component,
i.e. `core` and `dram` are minus background
and `base` is plus background of `core` and `dram`.
}
\examples{
e <- c("base","core","dram")
ret <- Insertionsort(runif(1000))
rbind(
  rawperf(ret)
, bacperf(ret)
, difperf(ret)
, adjperf(ret)
)[,e]
all.equal(sum(rawperf(ret)[,e]), sum(adjperf(ret)[,e]))
}
\seealso{
\code{\link{rawperf}},  \code{\link{bacperf}}, \code{\link{difperf}}, \code{\link{adjperf}}, \code{\link{amdperf}}, \code{\link{perf}}
}
