% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{aEnergy}
\alias{aEnergy}
\title{Extract all Energy measurements}
\usage{
aEnergy(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a matrix with different measurements in columns
}
\description{
Extract all Energy measurements
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
aEnergy(ret)
}
