% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{amdperf}
\alias{amdperf}
\title{Extract adjusted energy cost for AMD}
\usage{
amdperf(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
measurements adjusted for background in the form of \code{\link{retperf}}
}
\description{
The energy values are adjusted such that
the `base` component is added to the `core` component
and `base` is set to zero, in measurement and baseline
then the same adjustment of  \code{\link{adjperf}} is done.
}
\examples{
e <- c("base","core","dram")
ret <- Insertionsort(runif(1000))
rbind(
  rawperf(ret)
, bacperf(ret)
, difperf(ret)
, adjperf(ret)
, amdperf(ret)
)[,e]
all.equal(sum(rawperf(ret)[,e]), sum(adjperf(ret)[,e]))
}
\seealso{
\code{\link{rawperf}},  \code{\link{bacperf}}, \code{\link{difperf}}, \code{\link{adjperf}}, \code{\link{amdperf}}, \code{\link{perf}}
}
