% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{aperf}
\alias{aperf}
\title{Extract all measurements}
\usage{
aperf(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a matrix with different measurements in columns
}
\description{
Extract all measurements
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
aperf(ret)
}
