% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{approxMedian}
\alias{approxMedian}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'aMd'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
approxMedian(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a double vector for searching (re-arranged with situation='insitu')}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
the approximate median as a double scalar with an attribute \code{\link{perf}} like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
y <- x[]; m <- approxMedian(y)
x
y
median(x)
m
}
\seealso{
\code{\link{median}} for R's standard median, \code{\link{Pickselect}} for an reliable exact P&P selection algorithm using it
}
