% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{aTime}
\alias{aTime}
\title{Extract all time measurements}
\usage{
aTime(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a matrix with different measurements in columns
}
\description{
Extract all time measurements
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
aTime(ret)
}
