% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{bacperf}
\alias{bacperf}
\title{Extract background energy cost estimation}
\usage{
bacperf(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
raw measurements, see \code{\link{retperf}}
}
\description{
The energy values are estimates of `.perfbackground`
created by \code{\link{perfcalibrate}}
}
\examples{
ret <- Insertionsort(runif(1000))
rbind(
  rawperf(ret)
, bacperf(ret)
)[,c("base","core","dram")]
}
\seealso{
\code{\link{rawperf}},  \code{\link{bacperf}}, \code{\link{difperf}}, \code{\link{adjperf}}, \code{\link{amdperf}}, \code{\link{perf}}
}
