% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{base}
\alias{base}
\alias{bEnergy}
\title{Extract bEnergy}
\usage{
base(x)

bEnergy(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'base' or 'bEnergy'
}
\description{
Returns the baseline Energy in Joule
}
\details{
\code{\link{base}} is the RAPL packet energy minus the core energy,
\code{\link{bEnergy}} is the same *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}
}
\section{Functions}{
\itemize{
\item \code{bEnergy()}: the baseline-adjusted equivalent to `base`

}}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
base(ret)
bEnergy(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aEnergy}}
}
