% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{bcdEnergy}
\alias{bcdEnergy}
\title{Extract bcdEnergy}
\usage{
bcdEnergy(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'bcdEnergy'
}
\description{
Returns the RAPL BASE + CORE + DRAM Energy  in Joule
}
\details{
\code{\link{bcdEnergy}} is \code{\link{bEnergy}} + \code{\link{cEnergy}} + \code{\link{dEnergy}}.
This is *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
bEnergy(ret)
cEnergy(ret)
dEnergy(ret)
bcdEnergy(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aEnergy}}, \code{\link{pcdEnergy}}, \code{\link{bcdFootprint}}
}
