% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Bimesort}
\alias{Bimesort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Bime'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Bimesort(x, situation = c("insitu", "exsitu"), method = c("pointer", "index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(BimesortB(y))}   # o-merging, 1 loop check
,{y <- x[]; sperf(Bimesort(y))}    # o-merging, 1 loop check
,{y <- x[]; sperf(Nocosort(y))}    # p-merging, 3 loop checks
,{y <- x[]; sperf(Knuthsort(y))}   # p-merging, 1 loop check
)
}
\seealso{
\code{\link{Nocosort}} for Sedgewick's nocopy mergesort with three loop checks
\code{\link{Knuthsort}} for a ping-pong algorithm with Knuth's merge (also one loop check)
\code{\link{BimesortB}} for a version tuned to avoid branch-misprediction
}
