% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{cdFootprint}
\alias{cdFootprint}
\title{Extract cdFootprint}
\usage{
cdFootprint(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'cdFootprint'
}
\description{
Returns the RAPL CORE + DRAM Footprint
}
\details{
\code{\link{cdFootprint}} is \code{\link{cdEnergy}} * \code{\link{\%RAM}}.
This is *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
"\%RAM"(ret)
cdEnergy(ret)
cdFootprint(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aFootprint}}, \code{\link{cdEnergy}}, \code{\link{bcdFootprint}}
}
