% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Chunksort}
\alias{Chunksort}
\title{Chunksort}
\usage{
Chunksort(
  x,
  b = sqrt(length(x)),
  direction = 1L,
  situation = c("insitu", "exsitu"),
  perf = TRUE
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{b}{an integer block size (default sqrt(N))}

\item{direction}{an integer: -1 is descending, 1 is ascending, 0 is ascending-descending}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{perf}{FALSE to avoid measurement at startup (when grants might still be missing)}
}
\value{
like \code{\link{rawperf}}
}
\description{
a utility for test data generation which sorts sqrt(N) chunks of sqrt(N) elements using \code{\link{Quicksort2}}
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
perf(Chunksort(x))
oldpar <- par(pty="s")
plot(x)
lines(x)
par(oldpar)
}
\seealso{
\code{\link{order}} for standard R ordering, \code{\link{Insertionsort}} for the sorting version,
}
