% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Copysort}
\alias{Copysort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Copy'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Copysort(x, situation = c("insitu", "exsitu"), method = c("pointer", "index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Copysort(y))}    # extra copy pass, 1 loop check
,{y <- x[]; sperf(Nocosort(y))}    # p-merging, 3 loop checks
,{y <- x[]; sperf(Simplsort(y))}   # p-merging, 2 loop checks, inline comparison
,{y <- x[]; sperf(Bimesort(y))}    # for Sedgewick's 'bitonic Mergesort' with one loop check
,{y <- x[]; sperf(Knuthsort(y))}   # for a version with Knuth's merge (one loop check)
)
}
\seealso{
\code{\link{Nocosort}} for Sedgewick's 'no copy merge sort' that avoid one copy pass per merge using the ping-pong optmization (but three loop checks)
\code{\link{Simplsort}} simple mergesort with 1 move pass and 2 loop checks (N elements buffer)
\code{\link{Bimesort}} for Sedgewicks 'bitonic' mergesort with single move pass, sentinel and single loop check (N elements buffer) but stable
\code{\link{Knuthsort}} for a ping-pong algorithm with Knuth's merge (one loop check)
}
