% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{core}
\alias{core}
\alias{cEnergy}
\title{Extract cEnergy}
\usage{
core(x)

cEnergy(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'core' or 'cEnergy'
}
\description{
Returns the RAPL core Energy in Joule
}
\details{
\code{\link{core}} is the RAPL core energy,
\code{\link{cEnergy}} is the same *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}
}
\section{Functions}{
\itemize{
\item \code{cEnergy()}: the baseline-adjusted equivalent to `core`

}}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
core(ret)
cEnergy(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aEnergy}}
}
