% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{cpuTime}
\alias{cpuTime}
\title{Extract cpuTime}
\usage{
cpuTime(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'cpuTime'
}
\description{
Returns the cpuTime in seconds.
}
\details{
Note that cpuTime is not measured but estimated
as the \code{\link{runTime}} multiplied by the estimated effective number of threads,
assuming no thread waits up to the number of physical cores.
}
\examples{
ret <- PKnuthsort(runif(1e6), threads=2)
perf(ret)
secs(ret)
runTime(ret)
cpuTime(ret)
}
\seealso{
\code{\link{perf}}
}
