% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{cran_compatible}
\alias{cran_compatible}
\title{cran_compatible}
\usage{
cran_compatible()
}
\value{
\code{TRUE} if the package has beed compiled under R's memory management (Calloc,Free) for CRAN conpatibility,
and returns \code{FALSE} if the package has beed compiled for with the standard C memory management (malloc,free).
}
\description{
Show compilation choice for memory management and random number generation
}
\examples{
  cran_compatible()
}
\seealso{
\code{\link{greeNsort-package}}
}
