% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Crocosort}
\alias{Crocosort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Croco'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Crocosort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Knuthsort(y))}
,{y <- x[]; sperf(TKnuthsort(y))}
,{y <- x[]; sperf(Crocosort(y))}
,{y <- x[]; sperf(Croco3sort(y))}
,{y <- x[]; sperf(Croco4sort(y))}
)
}
\seealso{
\code{\link{Knuthsort}} for the classic ping-pong algorithm with distant buffer
\code{\link{TKnuthsort}} for the algorithm with relocation-moves of both input-streams (and Knuth's merge)
\code{\link{Crocosort}} for a r-merging algorithm using relocation moves of only one input stream (and semi-inplace merging)
\code{\link{Croco3sort}} and \code{\link{Croco4sort}} for k-ary versions with partial-inplace merging
}
