% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{current_compare}
\alias{current_compare}
\title{string compare}
\usage{
current_compare(x, y)
}
\arguments{
\item{x}{a vector of \code{\link{character}}}

\item{y}{a vector of \code{\link{character}}}
}
\value{
-1,0,+1 for x<y,x==y,x>y
}
\description{
access to the 3-way string compare function compiled into the size-varying (string) sorting routines
}
\examples{
current_compare("1.1","1.1")
current_compare("1.1","1.9")
current_compare("1.9","1.1")
current_compare(1.1,1.1)
current_compare(1.1,1.9)
current_compare(1.9,1.1)
}
