% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{current_issorted}
\alias{current_issorted}
\title{check sortedness of string}
\usage{
current_issorted(x)
}
\arguments{
\item{x}{a vector of \code{\link{double}} or \code{\link{character}}}
}
\value{
logical
}
\description{
check sortedness of string given the definition used in \code{\link{current_compare}}
}
\examples{
current_issorted(c("1.1","1.1"))
current_issorted(c("1.1","1.9"))
current_issorted(c("1.9","1.1"))
current_issorted(c(1.1,1.1))
current_issorted(c(1.1,1.9))
current_issorted(c(1.9,1.1))
current_issorted(c("a","\u00bbWas","z.b","z.a"))
!is.unsorted(c("a","\u00bbWas","z.b","z.a"))
current_issorted(c(1.9, 1.1, 9.9, 9.1))
!is.unsorted(c(1.9, 1.1, 9.9, 9.1))
}
