% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{current_keys}
\alias{current_keys}
\title{current_keys}
\usage{
current_keys(x)
}
\arguments{
\item{x}{a double or character vector to be rounded (or not)}
}
\value{
The rounded or not-rounded input values
}
\description{
Show compilation choice for stability testing
}
\details{
The package allows rounding of key values before comparision for testing of sorting stability:
the sorting comparator does not see decimals
and the sporting algorithm must not change the order of decimals within ties.
This provides an API into the rounding (or not rounding) of the key values.
}
\examples{
  current_keys(seq(-10, 10, 0.5))
  current_keys(num2str(seq(-10, 10, 0.5)))
}
\seealso{
\code{\link{greeNsort-package}}
}
