% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{debias}
\alias{debias}
\title{Debias measurement}
\usage{
debias(
  measurement = NULL,
  sleep = measurement$sleep,
  task = measurement$task,
  with.background = FALSE,
  used.cores = 1,
  available.cores = perfcores(FALSE)
)
}
\arguments{
\item{measurement}{either a list with two components \code{sleep} and \code{task} or the following two parameters}

\item{sleep}{background-costs: a vector of measurements from \code{\link{measure1}} or a matrix of measurements from \code{\link{measures}}}

\item{task}{algorithm-costs: a vector of measurements from \code{\link{measure1}} or a matrix of measurements from \code{\link{measures}}}

\item{with.background}{TRUE to add back 1/p of the sleep costs during runtime (default FALSE)}

\item{used.cores}{number of cores used (default 1)}

\item{available.cores}{number of cores available (default number of physical cores, see \code{\link{perfcores}})}
}
\value{
the debiased task input (without \code{seconds-elapsed})
}
\description{
Removes sleep-cost from task-cost
and optionally assigns 1/p of proportional sleep costs during runtime
where p is the number of physical cores (not hyperthreaded cores)
}
\note{
that this measurement and correction requires relatively large sorting tasks
}
\examples{
\donttest{
.sudopw <- rstudioapi::askForPassword("sudopw")
m <- measure("Frog1","tieboot",2^22,5,.sudopw,".perf.txt")
d <- debias(m)
rbind(
  s=apply(m$sleep,2,median)
, t=apply(m$task,2,median)
, d=apply(d,2,median)
)
rm(.sudopw)
}
}
