% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{debiasUVW}
\alias{debiasUVW}
\title{Debias varying (UWV) measurement}
\usage{
debiasUVW(measurement, overhead)
}
\arguments{
\item{measurement}{a vector of measurements from \code{\link{measure1}} or a matrix of measurements from \code{\link{measures}} of a size varying algorithm}

\item{overhead}{a vector of measurements from \code{\link{measure1}} or a matrix of measurements from \code{\link{measures}} of \code{\link{UVWoverhead}}}
}
\value{
the debiased measurement input
}
\description{
Removes overhead costs from measuerment costs (except for the mesaurements from \code{\link{perf}})
}
\note{
that this measurement and correction requires relatively large sorting tasks
}
\examples{
\donttest{
.sudopw <- rstudioapi::askForPassword("sudopw")
m <- debias(measure("WQuick2","tieboot",2^22,3,.sudopw,".perf.txt"))
o <- debias(measure("UVW","tieboot",2^22,3,.sudopw,".perf.txt"))
d <- debiasUVW(m,o)
rbind(
  m=apply(m,2,median)
, o=apply(o,2,median)
, d=apply(d,2,median)
)
rm(.sudopw)
}
}
