% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{DietCPTsort}
\alias{DietCPTsort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'CPT'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
DietCPTsort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(DietACPTsort(y))}
,{y <- x[]; sperf(DietCPTsort(y))}
,{y <- x[]; sperf(DietCPsort(y))}
,{y <- x[]; sperf(Swansort(y))}
,{y <- x[]; sperf(DietCPsort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(Swansort(y, situation = 'exsitu'))}
)
x <- testdb$func$tielog2(n)
rbind(
 {y <- x[]; sperf(DietACPTsort(y))}
,{y <- x[]; sperf(DietCPTsort(y))}
,{y <- x[]; sperf(DietCPsort(y))}
,{y <- x[]; sperf(Swansort(y))}
,{y <- x[]; sperf(DietCPsort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(Swansort(y, situation = 'exsitu'))}
)
}
\seealso{
\code{\link{DietACPTsort}} for a t-partition algorithm using \code{\link{approxMedian}} as pivot.
\code{\link{DietCPsort}} for a ping-pong algorithm without relocation moves.
\code{\link{Swansort}} for a faster b-partition algorithm saving the extra counting pass by wing-partitioning
}
