% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{difperf}
\alias{difperf}
\title{Extract energy cost minus background}
\usage{
difperf(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
measurements minus background in the form of \code{\link{retperf}}
}
\description{
The energy values are the difference between
raw measurements (\code{\link{rawperf}}) and
background estimates (\code{\link{bacperf}})
}
\examples{
e <- c("base","core","dram")
ret <- Insertionsort(runif(1000))
rbind(
  rawperf(ret) - bacperf(ret)
, difperf(ret)
)[,e]
all.equal(sum(rawperf(ret)[,e]), sum(difperf(ret)[,e]))
}
\seealso{
\code{\link{rawperf}},  \code{\link{bacperf}}, \code{\link{difperf}}, \code{\link{adjperf}}, \code{\link{amdperf}}, \code{\link{perf}}
}
