% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{dram}
\alias{dram}
\alias{dEnergy}
\title{Extract dEnergy}
\usage{
dram(x)

dEnergy(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'dram' or 'dEnergy'
}
\description{
Returns the RAPL DRAM in Joule
}
\details{
\code{\link{dram}} is the RAPL DRAM energy,
\code{\link{dEnergy}} is the same *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}
}
\section{Functions}{
\itemize{
\item \code{dEnergy()}: the baseline-adjusted equivalent to `dram`

}}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
dram(ret)
dEnergy(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aEnergy}}
}
