% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{DucksortB}
\alias{DucksortB}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'DuckB'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
DucksortB(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Zucksort(y))}
,{y <- x[]; sperf(ZucksortA(y))}
,{y <- x[]; sperf(Ducksort(y))}
,{y <- x[]; sperf(DucksortB(y))}
)
x <- testdb$func$tiesqrt(n)
rbind(
 {y <- x[]; sperf(Zucksort(y))}
,{y <- x[]; sperf(ZucksortA(y))}
,{y <- x[]; sperf(Ducksort(y))}
,{y <- x[]; sperf(DucksortB(y))}
)
x <- testdb$func$ascall(n)
rbind(
 {y <- x[]; sperf(Zucksort(y))}
,{y <- x[]; sperf(ZucksortA(y))}
,{y <- x[]; sperf(Ducksort(y))}
,{y <- x[]; sperf(DucksortB(y))}
)
}
\seealso{
\code{\link{Ducksort}} for the non-block-tuned version
\code{\link{Zucksort}} the greeNsort algorithm for early termination on ties
\code{\link{ZucksortA}} for the simple asymmetric basis for this
\code{\link{Pdqsort}} for Patterm Defeating Quicksort (branchless)
\code{\link{PDucksortB}} for a parallel version
}
