% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogsort0}
\alias{Frogsort0}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog0'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogsort0(x, situation = c("insitu", "exsitu"), method = c("index", "pointer"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort0(y))}
,{y <- x[]; sperf(Geckosort0(y))}  # a version with symmetric adaptivity
,{y <- x[]; sperf(Frogsort1(y))}   # a version that handles odd and even numbers of elements
)
}
\seealso{
\code{\link{Frogsort1}} for a more general version that handles odd and even numbers of elements without a dummy element
\code{\link{Geckosort0}} for a version with symmetric adaptivity
\code{\link{PFrogsort0}} for a more general parallel version that special cases incomplete triplets
}
