% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogsort1A}
\alias{Frogsort1A}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog1A'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogsort1A(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer", "index")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$ascnorm(n)
rbind(
 {y <- x[]; sperf(Frogsort1(y))}   # the untuned balanced version
,{y <- x[]; sperf(Frogsort1A(y))}
)
}
\seealso{
\code{\link{Knuthsort}} for p-merging with 100\% buffer
\code{\link{KnuthsortA}} for p-merging with 100\% buffer
\code{\link{Omitsort}} for a p-merging with 100\% buffer tuned for adaptivity
\code{\link{Frogsort1A}} for a version tuned with non-overlap testing for adaptivity on presorted data
\code{\link{Squidsort1}} for a version with adaptivity for presorted and reverse-sorted
\code{\link{VFrogsort1A}} for a version sorting strings (directly)
}
