% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogsort1B}
\alias{Frogsort1B}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog1B'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogsort1B(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer", "index")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort1B(y))}
,{y <- x[]; sperf(Frogsort1(y))}   # the basic balanced version
,{y <- x[]; sperf(KatasortB(y))}   # a tuned p-merging with 100\\% buffer
,{y <- x[]; sperf(BimesortB(y))}   # a tuned o-merging with 100\\% buffer
,{y <- x[]; sperf(Simplsort(y))}   # another p-merging with 100\\% buffer
,{y <- x[]; sperf(Frogsort1A(y))}  # a version tuned for adaptivity
)
}
\seealso{
\code{\link{Frogsort1}}  for the basic balanced version
\code{\link{KatasortB}}  for a tuned p-merging with 100\% buffer
\code{\link{BimesortB}}  for a tuned p-merging with 100\% buffer
\code{\link{Simplsort}}  for a p-merging with 100\% buffer that also improves on B
\code{\link{Frogsort1A}} for a version tuned for adaptivity
}
