% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogsort2}
\alias{Frogsort2}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog2'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogsort2(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer", "index"),
  p = 1/7
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}

\item{p}{fraction of  buffer relative to data size}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort2(y))}   # the unbalanced version using F&M
,{y <- x[]; sperf(Frogsort1(y))}   # the basic balanced version
,{y <- x[]; sperf(Frogsort1B(y))}  # a balanced version tuned to avoid branch-(mis)prediction
,{y <- x[]; sperf(Frogsort3(y))}   # the unbalanced version using A&M
,{y <- x[]; sperf(Frogsort6(y))}   # a parametrized generalization of Frogsort1,2,3
)
}
\seealso{
\code{\link{Frogsort1}}  for the basic balanced version
\code{\link{Frogsort2A}} for a version tuned for presorting
\code{\link{Frogsort2B}} for a version tuned against branch-misprediction
\code{\link{Frogsort3}}  for an unbalanced version using A&M requiring even less buffer
\code{\link{Frogsort6}}  for a parametrized generalization of Frogsort1,2,3
\code{\link{MFrogsort2}}  for a version sorting matrix columns by keys in first row (moves keys and payload)
\code{\link{NFrogsort2}}  for a version sorting matrix columns by keys in first row (moves keys and pointers)
\code{\link{PFrogsort2}}  for a parallel version
}
