% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogsort3B}
\alias{Frogsort3B}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog3B'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogsort3B(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer", "index"),
  p = 1/8
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}

\item{p}{fraction of  buffer relative to data size}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort3(y))}    # the untuned version
,{y <- x[]; sperf(Frogsort1B(y))}   # the balanced version
,{y <- x[]; sperf(Frogsort2B(y))}   # the unbalanced version using F&M
,{y <- x[]; sperf(Frogsort3B(y))}   # the unbalanced version using A&M
)
}
\seealso{
\code{\link{Frogsort3}} for the untuned version
\code{\link{Frogsort1B}} for the underlying balanced version using triplets
\code{\link{Frogsort2B}} for an unbalanced version using F&M requiring even less buffer and often even faster
}
