% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogsort4}
\alias{Frogsort4}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog4'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogsort4(
  x,
  situation = c("insitu", "exsitu"),
  method = c("index"),
  p = 1/sqrt(length(x)),
  verbose = FALSE
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}

\item{p}{fraction of initial buffer relative to data size}

\item{verbose}{TRUE reports buffer fraction p, absolute buffer size b and data size n}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort1(y))}
,{y <- x[]; sperf(Frogsort4(y))}
,{y <- x[]; sperf(Frogsort5(y))}
,{y <- x[]; sperf(Jumpsort(y))}
)
}
\seealso{
\code{\link{Frogsort5}} for a bottom-up version with optimized (bigger) buffer size delaying N/2 buffer need
\code{\link{Frogsort1}} for the top-down version with 50\% buffer
\code{\link{Jumpsort}}  for an algorithm with sqrt(N) buffer all the time, also distance-reducing
}
