% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Frogtop5}
\alias{Frogtop5}
\alias{Frogbot5}
\alias{Frogbuf5}
\alias{Frogsort5}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Frog5'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Frogtop5(n, situation = c("insitu", "exsitu"))

Frogbot5(n, situation = c("insitu", "exsitu"))

Frogbuf5(n, situation = c("insitu", "exsitu"))

Frogsort5(
  x,
  situation = c("insitu", "exsitu"),
  method = c("index"),
  p = Frogbuf5(length(x), situation = situation),
  verbose = FALSE
)
}
\arguments{
\item{n}{size of the vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{x}{a \code{\link{double}} vector to be sorted}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}

\item{p}{fraction of initial buffer relative to data size}

\item{verbose}{TRUE reports buffer fraction p, absolute buffer size b and data size n}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\section{Functions}{
\itemize{
\item \code{Frogtop5()}: calculate the footprint-optimal number of top-levels in the binary merge tree that use N/2 buffer

\item \code{Frogbot5()}: calculate the footprint-optimal number of bottom-levels in the binary merge tree that use the initial buffer

\item \code{Frogbuf5()}: calculate the footprint-optimal buffer-fraction

}}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort1(y))}
,{y <- x[]; sperf(Frogsort4(y))}
,{y <- x[]; sperf(Frogsort5(y))}
,{y <- x[]; sperf(Jumpsort(y))}
)
}
\seealso{
\code{\link{Frogsort1}} for the top-down version with 50\% buffer
\code{\link{Frogsort4}} for the bottom-up version with sqrt(N) initial buffer size
\code{\link{Jumpsort}}  for an algorithm with sqrt(N) buffer all the time, also distance-reducing
}
