% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Geckosort1}
\alias{Geckosort1}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Gecko1'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Geckosort1(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer", "index")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Geckosort1(y))}
,{y <- x[]; sperf(Geckosort0(y))}   # the root version using triplets
,{y <- x[]; sperf(Frogsort1(y))}    # a version with asymmetric adaptivity
)
}
\seealso{
\code{\link{Geckosort0}} for the basic version operating on triplets
\code{\link{Frogsort1}} for the symmetrically adaptive version
}
