% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Insertionorder}
\alias{Insertionorder}
\title{Insertionorder}
\usage{
Insertionorder(
  x,
  o = seq_along(x),
  direction = c("l2r", "r2l", "rev"),
  situation = c("insitu", "exsitu")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{o}{a vector with integer positions in x to be sorted with x}

\item{direction}{"l2r" for reading and writing stable ties from left to right, "r2l" or "rev" for writing (hence reversing) stable ties from right to left, i.e. the result is reverse ascening or reverse descending.}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}
}
\value{
like \code{\link{rawperf}}
}
\description{
like \code{\link{Insertionsort}} but sorting an index instead of the data (random access to the data)
}
\examples{
x <- testdb$func$permut(2^6)
i <- seq_along(x)
sperf(Insertionorder(x, i))
x
x[i]
}
\seealso{
\code{\link{order}} for standard R ordering, \code{\link{Insertionsort}} for the sorting version,
}
