% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Insertionsort}
\alias{Insertionsort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Ins'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Insertionsort(
  x,
  direction = c("l2r", "r2l", "rev"),
  decreasing = FALSE,
  situation = c("insitu", "exsitu"),
  method = c("index")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{direction}{"l2r" for reading and writing stable ties from left to right, "r2l" or "rev" for writing (hence reversing) stable ties from right to left, i.e. the result is reverse ascening or reverse descending.}

\item{decreasing}{default \code{FALSE} for ascending,  \code{TRUE} for descending}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
x <- testdb$func$permut(2^6)
sperf(Insertionsort(x))
x
}
\seealso{
\code{\link{sort}} for standard R sorting,
\code{\link{Selectionsort}} for Selectionsort,
\code{\link{RInsertionsort}} and \code{\link{SInsertionsort}} for indirect versions,
\code{\link{UInsertionsort}} and \code{\link{WInsertionsort}} for indirect size-varying versions,
and \code{\link{VInsertionsort}} for a direct size-varying version.
Finally  \code{\link{MInsertionsort}} for sorting of matrix-columns by first row-values
and \code{\link{NInsertionsort}} for sorting such matrices by sorting separated keys and then permuting the columns.
}
