% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{insertionsort_limit}
\alias{insertionsort_limit}
\title{insertionsort_limit}
\usage{
insertionsort_limit()
}
\value{
the max value for which recursion is stopped and sorting is delegated to \code{\link{Insertionsort}}.
}
\description{
Show compilation choice for insertion sort tuning
}
\details{
The package allows compiling with different choices for insertion sort tuning,
namely the max value for which recursion is stopped and sorting is delegated to \code{\link{Insertionsort}}.
A value of 0 has a special meaning (may compile to somewhat different code for the no-tuning-case).
}
\examples{
  insertionsort_limit()
}
\seealso{
\code{\link{greeNsort-package}}
}
