% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Katasort}
\alias{Katasort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Kata'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Katasort(x, situation = c("insitu", "exsitu"), method = c("pointer", "index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Knuthsort(y))}   # p-merging, 1 loop check
,{y <- x[]; sperf(Katasort(y))}    # p-merging, tuned for 1/2 loop check
,{y <- x[]; sperf(KatasortA(y))}   # p-merging, tuned for 1/2 loop check, tuned A
,{y <- x[]; sperf(KatasortB(y))}   # p-merging, tuned for 1/2 loop check, tuned B
,{y <- x[]; sperf(Kata3sort(y))}   # p-merging, tuned for 1/2 loop check, log3 S&M
,{y <- x[]; sperf(Kata4sort(y))}   # p-merging, tuned for 1/2 loop check, log4 S&M
)
}
\seealso{
\code{\link{Knuthsort}} for a version with Knuth's merge (one check per loop traversal)
\code{\link{KatasortA}} for a version tuned for less initial copying and non-overlap check
\code{\link{KatasortB}} for a version tuned to avoid branch-misprediction
\code{\link{Kata3sort}} and \code{\link{Kata4sort}} for k-ary versions
}
