% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{KatasortA}
\alias{KatasortA}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'KataA'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
KatasortA(x, situation = c("insitu", "exsitu"), method = c("pointer", "index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$ascnorm(n)
rbind(
 {y <- x[]; sperf(Katasort(y))}    # p-merging, 1/2 loop check
,{y <- x[]; sperf(KatasortA(y))}   # p-merging, 1/2 loop check, tuned A
,{y <- x[]; sperf(KnuthsortA(y))}  # p-merging, 1 loop check, tuned A
,{y <- x[]; sperf(Frogsort1A(y))}  # s-merging algorithm, 50\\% buffer, 1 loop check, tuned A
)
}
\seealso{
\code{\link{Katasort}} for the version without this tuning (half loop check)
\code{\link{KnuthsortA}} for an algorithm with Knuth's merge with this tuning (one loop check)
\code{\link{Frogsort1A}} for a s-merging algorithm with non-overlap check that requries only 50\% buffer (one loop check)
}
