% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Kiwisort}
\alias{Kiwisort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Kiwi'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Kiwisort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(DietCPsort(y))}
,{y <- x[]; sperf(DietPcsort(y))}
,{y <- x[]; sperf(Kiwisort(y))}
,{y <- x[]; sperf(Swansort(y))}
,{y <- x[]; sperf(Swansort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(Kiwisort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortA(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortB(y, situation = 'exsitu'))}
,{y <- x[]; sperf(SwansortB(y, situation = 'exsitu'))}
)
x <- testdb$func$tielog2(n)
rbind(
 {y <- x[]; sperf(DietCPsort(y))}
,{y <- x[]; sperf(DietPcsort(y))}
,{y <- x[]; sperf(Kiwisort(y))}
,{y <- x[]; sperf(Swansort(y))}
,{y <- x[]; sperf(Swansort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(Kiwisort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortA(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortB(y, situation = 'exsitu'))}
,{y <- x[]; sperf(SwansortB(y, situation = 'exsitu'))}
)
}
\seealso{
\code{\link{DietCPsort}} for a ping-pong algorithm with an extra counting pass,
\code{\link{DietPcsort}} for a ping-pong algorithm which also tries to avoid an extra counting pass (less successfully),
\code{\link{KiwisortA}} for a version tuned for presorted data
\code{\link{KiwisortB}} for a version tuned to avoid branch-(mis)prediction.
\code{\link{Swansort}} for a b-partition algorithm (slower in 'insitu' setting)
\code{\link{SwansortB}} for a b-partition algorithm block-tuned (slower in 'insitu' setting)
}
