% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{KiwisortB}
\alias{KiwisortB}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'KiwiB'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
KiwisortB(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Kiwisort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortA(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortB(y, situation = 'exsitu'))}
,{y <- x[]; sperf(SwansortB(y, situation = 'exsitu'))}
)
x <- testdb$func$ascall(n)
rbind(
 {y <- x[]; sperf(Kiwisort(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortA(y, situation = 'exsitu'))}
,{y <- x[]; sperf(KiwisortB(y, situation = 'exsitu'))}
,{y <- x[]; sperf(SwansortB(y, situation = 'exsitu'))}
)
}
\seealso{
\code{\link{Kiwisort}} for the untuned version
\code{\link{KiwisortA}} for a version tuned for presorted data
\code{\link{SwansortB}} for a distance-reducing version tuned to avoid branch-(mis)prediction.
}
