% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Knuth3sort}
\alias{Knuth3sort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Knuth3'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Knuth3sort(x, situation = c("insitu", "exsitu"), method = c("pointer"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Knuthsort(y))}
,{y <- x[]; sperf(Knuth3sort(y))}
,{y <- x[]; sperf(Knuth4sort(y))}
,{y <- x[]; sperf(Kata3sort(y))}
,{y <- x[]; sperf(Croco3sort(y))}
)
}
\seealso{
\code{\link{Knuthsort}} for the binary version
\code{\link{Knuth4sort}} for a log4 version
\code{\link{Kata3sort}} for a version with Katajainen's tuned merge (on average half check per loop traversal)
\code{\link{Croco3sort}} for a r-merging algorithm that reduces distance
}
