% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{lifeTime}
\alias{lifeTime}
\title{Extract lifeTime}
\usage{
lifeTime(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'lifeTime'
}
\description{
Returns the lifeTime in seconds
}
\details{
lifeTime is \code{\link{runTime}} divided by the number of parallel processes that share the hardware for amortization
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
secs(ret)
runTime(ret)
}
\seealso{
\code{\link{perf}}
}
