% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measure1}
\alias{measure1}
\title{Measuring one algorithm in one data setting}
\usage{
measure1(pass, file, algo, data, size, repl = 1, expected = NULL, ...)
}
\arguments{
\item{pass}{the root password for \code{sudo}}

\item{file}{the name of the file into which \code{perf} is redirected}

\item{algo}{label of the algorithm as in \code{\link{algodb}$label}}

\item{data}{name of the test data pattern as in \code{\link{testdb}$name}}

\item{size}{number of test elements to be sorted}

\item{repl}{the replications number (for \code{\link{set.seed}})}

\item{expected}{the expected runtime in seconds}

\item{\dots}{further arguments passed to the algorithm}
}
\value{
a list with two components
\item{sleep}{the sleep measurement: a vector returned by \code{\link{measures}}}
\item{task}{the algorithm mesasurement: a vector returned by \code{\link{measures}}}
}
\description{
For each replication \code{r in 0:repl} test \code{data} sets is created with \code{set.seed(r)}.
Data number 0 is used to determine the expected runtime of \code{algo}.
For each of the remaining \code{1:repl} data sets \code{\link{measure1}} is called twice,
once measuring a sleep command of the expected duration
and once measuring algorithm \code{algo} on the data.
}
\note{
The recommended correction for background
If the algorithm is size-varying the data is coerced to integer and then to character.
Note that the size-varing algorithms return their pure sorting time
but have relevant overhead for pre- and post-processing the strings
which is included in this measurement,
to correct for this bias measure also \code{\link{UVWoverhead}}.
}
\examples{
\donttest{
.sudopw <- rstudioapi::askForPassword("sudopw")
n <- 2^23
a <- "Frog1"
d <- "tieboot"
m1 <- measure1(.sudopw, ".perf.txt", a, d, n, 1)
m2 <- measure1(.sudopw, ".perf.txt", a, d, n, 2, expected=m1$task["runtime"])
m1 <- measure1(.sudopw, ".perf.txt", "VFrog1", "tieboot", 2^23, 1)
m2 <- measure1(.sudopw, ".perf.txt", "VFrog1", "tieboot", 2^23, 2, expected=m1$task["runtime"])
rm(.sudopw)
}
}
