% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measurea}
\alias{measurea}
\title{Measuring one algorithm in one data setting}
\usage{
measurea(pass, file, algo, data, size, repl, ret = list(), imag = NULL, ...)
}
\arguments{
\item{pass}{the root password for \code{sudo}}

\item{file}{the name of the file into which \code{perf} is redirected}

\item{algo}{label of the algorithm as in \code{\link{algodb}$label}}

\item{data}{name of the test data pattern as in \code{\link{testdb}$name}}

\item{size}{number of test elements to be sorted}

\item{repl}{the replications id (only for reporting)}

\item{ret}{optional a list with partial results}

\item{imag}{NULL or a filename for saving the image}

\item{\dots}{further arguments passed to the algorithm}
}
\value{
a list with \code{repl} components
each with \code{data} components
each with \code{algo} components
each with two components
\item{sleep}{the sleep measurement: a vector returned by \code{\link{measures}}}
\item{task}{the algorithm mesasurement: a vector returned by \code{\link{measures}}}
}
\description{
For each replication \code{r in 0:repl} test \code{data} sets is created with \code{set.seed(r)}.
Data number 0 is used to determine the expected runtime of \code{algo}.
For each of the remaining \code{1:repl} data sets \code{\link{measure1}} is called.
}
\note{
The recommended correction for background
If the algorithm is size-varying the data is coerced to integer and then to character.
Note that the size-varing algorithms return their pure sorting time
but have relevant overhead for pre- and post-processing the strings
which is included in this measurement,
to correct for this bias measure also \code{\link{UVWoverhead}}.
}
\examples{
\donttest{
.sudopw <- rstudioapi::askForPassword("sudopw")
l <- measurea(.sudopw, ".perf.txt", "Frog1", "asclocal", 2^23, 2, imag="measurea.RData")
l <- measurea(.sudopw, ".perf.txt", c("Frog1", "Frog2"), c("asclocal", "ascglobal"), 2^23, 3, imag="measurea.RData", ret=l)
l <- measurea(.sudopw, ".perf.txt", "Ska", "tiesqrt", 2^24, 50)
rm(.sudopw)
}
}
