% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R
\name{measures}
\alias{measures}
\title{Measuring a single expresssion}
\usage{
measures(pass, file, expr, expected, offset = 0.2, slope = 1.1, wait = 0.1)
}
\arguments{
\item{pass}{the root password for \code{sudo}}

\item{file}{the name of the file into which \code{perf} is redirected}

\item{expr}{an expression calling a sorting function with output suitable to \code{\link{perf}} to be measured}

\item{expected}{the expected runtime in seconds}

\item{offset}{security offset in seconds}

\item{slope}{security factor}

\item{wait}{seconds waited between spawning \code{perf} and evaluating \code{expr} and also waited after the sleep time}
}
\value{
a vector with elements
\item{instructions}{instructions (count)}
\item{cpu-cycles}{cpu-cycles (count)}
\item{bus-cycles}{bus-cycles (count)}
\item{page-faults}{page-faults (count)}
\item{cache-references}{cache-references (count)}
\item{cache-misses}{cache-misses (count)}
\item{L1-dcache-loads}{L1-dcache-loads (count)}
\item{L1-dcache-stores}{L1-dcache-stores (count)}
\item{branch-instructions}{branch-instructions (count)}
\item{branch-misses}{branch-misses (count)}
\item{energy-pkg}{power/energy-pkg/ (Joule)}
\item{energy-cores}{power/energy-cores/ (Joule)}
\item{energy-ram}{power/energy-ram/ (Joule)}
\item{seconds-elapsed}{the time perf measured for sleep (seconds)}
\item{memory}{memory (\code{\link{size}})}
\item{runtime}{runtime (\code{\link{secs}})}
\item{footprint}{footprint (\code{\link{tFootprint}})}
}
\description{
This function spawns a seperated process which runs the linux \code{perf} tool and collects
interesting statistics including RAPL measurements
}
\details{
The sleep-time for linux perf is calculated as \code{offset + slope*expected}
}
\note{
\code{perf} requires
sudo apt-get install linux-tools-generic
sudo apt-get remove linux-tools-5.3.0-28-generic
sudo apt-get install linux-tools-5.3.0-40-generic
}
\examples{
\donttest{
.sudopw <- rstudioapi::askForPassword("sudopw")
x <- testdb$func$tieboot(2^23)
y <- x[]; s <- secs(Frogsort1(y))
y <- x[]; a <- measures(.sudopw, ".perf.txt", quote(Frogsort1(y)), s)
b <- measures(.sudopw, ".perf.txt", quote(Sys.sleep(s)), s)
d <- debias(sleep=b, task=a)
print(d)
rm(.sudopw)
}
}
