% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{MFrogsort2}
\alias{MFrogsort2}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'MFrog2'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
MFrogsort2(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer"),
  p = 1/7
)
}
\arguments{
\item{x}{a \code{\link{integer}} matrix to be sorted columnwise by key in the first row}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}

\item{p}{fraction of  buffer relative to data size}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
x <- as.integer(testdb$func$permut(2^4))
x <- rbind(x,-x,-x)
sperf(NInsertionsort(x))
print(x)
n <- getOption("greensort_example_size")
x <- as.integer(testdb$func$permut(n))
x <- rbind(x, -x, -x)
rbind(
 {y <- x[]; sperf(MKnuthsort(y))}
,{y <- x[]; sperf(NKnuthsort(y))}
,{y <- x[]; sperf(MFrogsort2(y))}
,{y <- x[]; sperf(NFrogsort2(y))}
)
}
\seealso{
\code{\link{Frogsort2}} for the standard algoritm for sorting a double vector
\code{\link{NFrogsort2}} for an algorithm sorting separated keys and pointers to records
\code{\link{MKnuthsort}} for a Knuthsort doing this
}
