% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{NInsertionsort}
\alias{NInsertionsort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'NIns'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
NInsertionsort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a integer matrix, columns to be sorted by the values in the first row}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
x <- as.integer(testdb$func$permut(2^4))
x <- rbind(x,-x,-x)
sperf(NInsertionsort(x))
x
}
\seealso{
\code{\link{Insertionsort}} for standard Insertionsort of vectors and
\code{\link{MInsertionsort}} for sorting of matrix-columns including non-separated keys
\code{\link{NKnuthsort}} for a scalable algorithm for sorting matrices using this
}
