% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Nocosort}
\alias{Nocosort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Noco'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Nocosort(x, situation = c("insitu", "exsitu"), method = c("pointer", "index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Nocosort(y))}    # p-merging, 3 loop checks
,{y <- x[]; sperf(Copysort(y))}    # extra copy pass, 1 loop check
,{y <- x[]; sperf(Simplsort(y))}   # p-merging, 2 loop checks, inline comparison
,{y <- x[]; sperf(Bimesort(y))}    # o-merging, 1 loop check
,{y <- x[]; sperf(Knuthsort(y))}   # p-merging, 1 loop check
,{y <- x[]; sperf(Katasort(y))}    # p-merging, tuned for 1/2 loop check
)
}
\seealso{
\code{\link{Copysort}} for a (faster) algorithm that does not user the ping-pong optmization
\code{\link{Bimesort}} for Sedgewick's 'bitonic Mergesort' with one loop check
\code{\link{Simplsort}} for a simple version with two loop checks
\code{\link{Knuthsort}} for a version with Knuth's merge (one loop check)
\code{\link{Katasort}} for a version with Katajainen's tuned merge (half loop check)
}
