% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntile2.R
\name{ntile2_asc_asc_right_to_left}
\alias{ntile2_asc_asc_right_to_left}
\title{ntiles of two ASC ASC sorted arrays from RIGHT to LEFT}
\usage{
ntile2_asc_asc_right_to_left(
  L,
  R,
  threads = 2,
  P = seq_len(threads)/threads,
  K = round(P * N)
)
}
\arguments{
\item{L}{left sorted vector (ascending)}

\item{R}{right sorted vector  (ascending)}

\item{threads}{scalar number of threads}

\item{P}{vector number of split fractions from the RIGHT (default: parameter 'threads' equally sized fractions)}

\item{K}{vector number of elements to be separated on the RIGHT (default: derived from parameter P)}
}
\value{
a 2xlength(K) matrix
}
\description{
ntiles of two ASC ASC sorted arrays from RIGHT to LEFT
}
\examples{
Ln = 2^12
Rn = 2^12
N = Ln+Rn
skew = runif(1, -2^8, 2^8)
L = rchisq(Ln, N-skew)
R = rchisq(Rn, N+skew)
L = L[order(current_keys(L))]
R = R[order(current_keys(R))]
threads <- 8
lr = ntile2_asc_asc_right_to_left(L,R,threads)
ylim = range(c(L,R))
lrna = lr; lrna[lrna==0] = NA
m = pmin(L[Ln+1-lrna["l",]], R[Rn+1-lrna["r",]], na.rm=TRUE)
oldpar <- par(mfrow=c(1,2), bg="grey40")
plot(L, ylim=ylim, col=rev(rep(seq_len(threads), diff(c(0,lr['l',])))), pch='.')
abline(h=m, col=seq_len(threads))
plot(R, ylim=ylim, col=rev(rep(seq_len(threads), diff(c(0,lr['r',])))), pch='.')
abline(h=m, col=seq_len(threads))
par(oldpar)
}
