% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntile2.R
\name{ntile2_asc_revasc_left_to_right}
\alias{ntile2_asc_revasc_left_to_right}
\title{ntiles of two ASC REVASC sorted arrays from LEFT to RIGHT}
\usage{
ntile2_asc_revasc_left_to_right(
  L,
  R,
  threads = 2,
  P = seq_len(threads)/threads,
  K = round(P * N)
)
}
\arguments{
\item{L}{left sorted vector (ascending)}

\item{R}{right sorted vector  (reverse ascending)}

\item{threads}{scalar number of threads}

\item{P}{vector number of split fractions from the LEFT (default: parameter 'threads' equally sized fractions)}

\item{K}{vector number of elements to be separated on the LEFT (default: derived from parameter P)}
}
\value{
a 2xlength(K) matrix
}
\description{
ntiles of two ASC REVASC sorted arrays from LEFT to RIGHT
}
\examples{
Ln = 2^12
Rn = 2^12
N = Ln+Rn
skew = runif(1, -2^8, 2^8)
L = rchisq(Ln, N-skew)
R = rchisq(Rn, N+skew)
L = L[order(current_keys(L))]
R = R[order(current_keys(R))]
threads <- 8
lr1 = ntile2_asc_asc_left_to_right(L,R,threads)
lr = ntile2_asc_revasc_left_to_right(L,rev(R),threads)
ylim = range(c(L,R))
lrna = lr; lrna[lrna==0] = NA
m = pmax(L[lrna["l",]], R[lrna["r",]], na.rm=TRUE)
oldpar <- par(mfrow=c(1,2), bg="grey40")
plot(L, ylim=ylim, col=rep(seq_len(threads), diff(c(0,lr['l',]))), pch='.')
abline(h=m, col=seq_len(threads))
plot(rev(R), ylim=ylim, col=rev(rep(seq_len(threads), diff(c(0,lr['r',])))), pch='.')
abline(h=m, col=seq_len(threads))
par(oldpar)
}
