% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{num2str}
\alias{num2str}
\title{Convert numeric test data to string test data}
\usage{
num2str(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a character vector
}
\description{
Convert numeric test data to string test data
}
\examples{
num2str(testdb$func$stable(12))
}
