% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Octosort}
\alias{Octosort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Octo'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Octosort(x, situation = c("insitu", "exsitu"), method = c("pointer"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$descnorm(n)
rbind(
 {y <- x[]; sperf(KnuthsortA(y))}
,{y <- x[]; sperf(Omitsort(y))}
,{y <- x[]; sperf(Octosort(y))}
,{y <- x[]; sperf(Frogsort1A(y))}
,{y <- x[]; sperf(Frogsort2A(y))}
,{y <- x[]; sperf(Squidsort1(y))}
,{y <- x[]; sperf(Squidsort2(y))}
,{y <- x[]; sperf(Timsort(y))}
,{y <- x[]; sperf(gfxTimsort(y))}
,{y <- x[]; sperf(Peeksort(y))}
,{y <- x[]; sperf(Powersort(y))}
,{y <- x[]; sperf(Powersort4(y))}
)
}
\seealso{
\code{\link{GKnuthsort}} for the non-tuned version
\code{\link{KnuthsortA}} for the tuned ping-pong algorithm
}
