% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Omitsort}
\alias{Omitsort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Omit'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Omitsort(x, situation = c("insitu", "exsitu"), method = c("pointer", "index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$ascnorm(n)
rbind(
 {y <- x[]; sperf(Knuthsort(y))}
,{y <- x[]; sperf(KnuthsortA(y))}
,{y <- x[]; sperf(Omitsort(y))}
,{y <- x[]; sperf(Frogsort1(y))}
,{y <- x[]; sperf(Frogsort1A(y))}
,{y <- x[]; sperf(Frogsort2(y))}
,{y <- x[]; sperf(Frogsort2A(y))}
,{y <- x[]; sperf(Timsort(y))}
,{y <- x[]; sperf(Insertionsort(y))}
)
}
\seealso{
\code{\link{Knuthsort}} for the version without this tuning (one loop check)
\code{\link{KnuthsortA}} for a version with simple tuning just skipping merging (but not copying)
\code{\link{Frogsort1}} for a s-merging algorithm without tuning that requires only 50\% buffer (one loop check)
\code{\link{Frogsort1A}} for a s-merging algorithm with non-overlap check that requires only 50\% buffer (one loop check)
\code{\link{Frogsort2}} for a memory-reduced s-merging algorithm that requires only 14\% buffer (one loop check)
\code{\link{Frogsort2A}} for a tuned memory-reduced s-merging algorithm that requires only 14\% buffer (one loop check)
\code{\link{Timsort}} for an algorithm that is tuned for the best case of presorted data
\code{\link{Insertionsort}} for an simple algorithm that performs very well on presorted data
}
