% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{optperf}
\alias{optperf}
\title{Automatically switch energy calculation}
\usage{
optperf(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
\code{\link{perf}}
}
\description{
Depending on \code{\link{getOption}} `greensort_perf_calc`
in `c('raw','dif','adj','amd')` extracts one of \code{\link{rawperf}},
\code{\link{difperf}} or \code{\link{adjperf}}
}
\examples{
getOption('greensort_perf_calc')
options(greensort_perf_calc='raw')
optperf(Insertionsort(runif(1000)))
}
\seealso{
\code{\link{bacperf}}, \code{\link{perfcalibrate}}, \code{\link{perf}}
}
