% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{pcdEnergy}
\alias{pcdEnergy}
\title{Extract pcdEnergy}
\usage{
pcdEnergy(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'pcdEnergy'
}
\description{
Returns the RAPL thread-partial BASE + CORE + DRAM in Joule
}
\details{
\code{\link{pcdEnergy}} is a part of \code{\link{bEnergy}} + \code{\link{cEnergy}} + \code{\link{dEnergy}}.
This is *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}#' see \code{\link{perfcalibrate}}
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
bEnergy(ret)
cEnergy(ret)
dEnergy(ret)
bcdEnergy(ret)
pcdEnergy(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aEnergy}}, \code{\link{bcdEnergy}}, \code{\link{pcdFootprint}}
}
