% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{pcdFootprint}
\alias{pcdFootprint}
\title{Extract pcdFootprint}
\usage{
pcdFootprint(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'pcdFootprint'
}
\description{
Returns the RAPL thread-partial BASE + CORE + DRAM Footprint
}
\details{
\code{\link{pcdFootprint}} is \code{\link{pcdEnergy}} * \code{\link{\%RAM}}.
This is *after* the adjustment specified in \code{\link{options}} `greensort_perf_calc` with default 'raw',
see \code{\link{perfcalibrate}}
}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
"\%RAM"(ret)
pcdEnergy(ret)
pcdFootprint(ret)
}
\seealso{
\code{\link{perf}}, \code{\link{aFootprint}}, \code{\link{pcdEnergy}}, \code{\link{bcdFootprint}}
}
