% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perf}
\alias{perf}
\alias{sperf}
\title{Performance measurements}
\usage{
perf(x)

sperf(x)
}
\arguments{
\item{x}{an object that has a `perf` \code{\link{attr}} or which is a perf attribute or which is a \code{\link{rbind}} \code{\link{matrix}} or \code{\link{data.frame}} of multiple such attributes}
}
\value{
a matrix or data.frame with \code{\link{rownames}} of the input and the following columns:
\item{\%RAM \code{\link{size}} or \code{\link{\%RAM}}}{the maximum memory used for execution (the size of data and buffer relative to the size of the data only)}
\item{runTime \code{\link{secs}} or \code{\link{runTime}}}{the execution time measured in seconds}
\item{lifeTime \code{\link{lifeTime}}}{the execution time measured in seconds corrected for parallel processes}
\item{cpuTime \code{\link{cpuTime}}}{the cpuTime estimated from runTime, number of threads and number of processes}
\item{bEnergy \code{\link{bEnergy}}}{the \code{\link{optperf}} corrected `base` energy }
\item{cdEnergy \code{\link{cdEnergy}}}{the \code{\link{optperf}} corrected `core` + `dram` energy }
\item{pcdEnergy \code{\link{pcdEnergy}}}{the \code{\link{optperf}} corrected thread adjusted partiel `base` + `core` + `dram` energy }
\item{bcdEnergy \code{\link{bcdEnergy}}}{the \code{\link{optperf}} corrected `base` + `core` + `dram` energy }
\item{cdFootprint \code{\link{cdFootprint}}}{ \code{\link{cdEnergy}} * \code{\link{\%RAM}} }
\item{pcdFootprint \code{\link{pcdFootprint}}}{ \code{\link{pcdEnergy}} * \code{\link{\%RAM}} }
\item{bcdFootprint \code{\link{bcdFootprint}}}{ \code{\link{bcdEnergy}} * \code{\link{\%RAM}} }
\item{tFootprint \code{\link{sizesecs}} or \code{\link{tFootprint}}}{the integral of memory size over execution time where size is measured as number of elements}
}
\description{
The `perf` function returns the default performance KPIs
}
\section{Functions}{
\itemize{
\item \code{sperf()}: a simple version of `perf` which returns only `\%RAM`, runTime`, `bcdEnergy`, `bcdFootprint`, `tFootprint`

}}
\examples{
N <- 1024
z <- Insertionsort(runif(N))
perf(z)
perf(rbind(z,z))
}
\seealso{
\code{\link{perfgrants}}, \code{\link{retperf}}, \code{\link{aTime}}, \code{\link{aEnergy}}, \code{\link{aFootprint}}, \code{\link{perfcalibrate}}
}
