% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfcalibrate}
\alias{perfcalibrate}
\title{Calibrate for background energy}
\usage{
perfcalibrate(n = 300, minsec = 0, maxsec = 24, plot = TRUE, save = FALSE)
}
\arguments{
\item{n}{number of measurements (length of \code{\link{seq}})}

\item{minsec}{minimum seconds to measure (begin of \code{\link{seq}})}

\item{maxsec}{maximum seconds to measure (end of \code{\link{seq}})}

\item{plot}{TRUE for plotting the models FALSE for not}

\item{save}{TRUE for saving .perfbackground in 'perfbackground.RData'}
}
\value{
\code{invisible} with side-effect in \code{\link{globalenv}} hidden environment \code{.perfbackground} with three components
\item{data}{a dataframe with columns of \code{\link{retperf}}}
\item{fits}{a list with three elements \code{base}, \code{core} and \code{dram} containing the linear models }
\item{pred}{a list with three elements \code{base}, \code{core} and \code{dram} containing the prediction functions }
}
\description{
Performs `n` sleep measures with \code{\link{perfsleep}},
fits linear models \code{\link{lm}} (without intercept) predicting energy-measurements
from (measured) sleep seconds.
}
\details{
The return object is saved to the \code{\link{globalenv}} under the name \code{.perfbackground}
}
\examples{
perfcalibrate(n=10, minsec=0, maxsec=2, plot = TRUE)
}
