% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfcheck}
\alias{perfcheck}
\title{Checking performance configuration}
\usage{
perfcheck()
}
\value{
an `TRUE` if the compiled comfiguration matches the machine, otherwise `FALSE`
}
\description{
Throws a warning if the powercap rapl configuration does not fit to the machine
}
\examples{
perfcheck()
}
\seealso{
\code{\link{perfmachine}},  \code{\link{perfcompiled}},  \code{\link{perfconfig}},  \code{\link{perf}}
}
