% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfcompiled}
\alias{perfcompiled}
\title{Compiled powercap paths}
\usage{
perfcompiled()
}
\value{
a character vector with three-elements
\item{package}{path to directory of PACKAGE counter}
\item{cores}{path to directory of CORES counter}
\item{dram}{path to directory of DRAM counter}
}
\description{
Return compiled paths to the powercap rapl directories
}
\examples{
perfcompiled()
}
\seealso{
\code{\link{perfmachine}},  \code{\link{perfcheck}},  \code{\link{perfconfig}},  \code{\link{perf}}
}
