% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfconfig}
\alias{perfconfig}
\title{Configure performance configuration}
\usage{
perfconfig()
}
\value{
path to the written config file
}
\description{
determine the paths for `package`, `core`, `dram` and overwrite the
C-header file which configures the paths to the powercap rapl directories
}
\details{
Call this in the package directory, then re-compile the package.
}
\examples{
\dontrun{
  perfconfig()
}
}
\seealso{
\code{\link{perf}}
}
