% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfcores}
\alias{perfcores}
\title{Default number of cores}
\usage{
perfcores(logical = TRUE)
}
\arguments{
\item{logical}{TRUE to get hyperthreads, false to get physical cores (at least under linux)}
}
\value{
integer number of cores
}
\description{
central interface defining the number of available cores
}
\examples{
perfcores()
perfcores(FALSE)
perfcores(TRUE)
#PKnuthsort(runif(), threads = perfcores(FALSE))

}
\seealso{
\code{\link{perf}}
}
