% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfgrants}
\alias{perfgrants}
\title{Granting access to performance counters}
\usage{
perfgrants()
}
\value{
invisible vector of grants
}
\description{
Determines the configured paths to the powercap rapl directories and executes the grants with `sudo`.
You need to know the sudo password.
}
\details{
After rebooting a machine the grants on the virtual powercap rapl directories are lost and need to be granted again.
This can either be done by root in a startup-script or here by calling `perfgrants`.
}
\examples{
\dontrun{
  perfgrants()
}
}
\seealso{
\code{\link{perf}}
}
