% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfmachine}
\alias{perfmachine}
\title{Powercap paths on this machine}
\usage{
perfmachine()
}
\value{
a character vector of checked paths with the discovered names attached
}
\description{
Return paths to the powercap rapl directories on this machine
}
\details{
Currently checks for four potential paths on a one-socket Intel or AMD machine.
Non-existing paths are returned as `NA`.
}
\examples{
perfmachine()
}
\seealso{
\code{\link{perfcompiled}},  \code{\link{perfcheck}},  \code{\link{perfconfig}},  \code{\link{perf}}
}
