% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfmax}
\alias{perfmax}
\title{Return performance maximum}
\usage{
perfmax()
}
\value{
a double vector with five elements
\item{time}{\code{\link{Inf}}}
\item{package}{PACKAGE maximum in Joule}
\item{core}{CORE maximum in Joule}
\item{uncore}{UNCORE maximum in Joule}
\item{dram}{DRAM maximum in Joule}
}
\description{
Return the maximum powercap RAPL counters
}
\examples{
perfmax()
}
\seealso{
\code{\link{perfnow}}, \code{\link{perfmon}}, \code{\link{perf}}
}
