% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfminus}
\alias{perfminus}
\title{Subtract (raw) performance measurements}
\usage{
perfminus(..., rowname = NULL)
}
\arguments{
\item{...}{one or more matrices of one or more rows of \code{\link{rawperf}}}

\item{rowname}{rowname for output, NULL means default, i.e. first rowname of input}
}
\value{
\code{\link{retperf}}
}
\description{
\code{\link{rbind}s} all arguments and aggregates all but the first them to a single row
and substracts the result from the first row.
Components \code{c("n","b","p","t")} are taken as is from the first row  and
component \code{\link{size}} is calculated like a  \code{\link{secs}} weighed average.
}
\examples{
x <- vtestdb$func$KafkaWords()
y <- x[]
p0 <- UVWoverhead(y)
p1 <- VKnuthsort(y)
p2 <- perfnow()
x <- sort(x)
p2 <- perfthen(p2, rowname="Rsort", size=1) # optimistic size
print(rbind(
  p0
, p1
, perfplus(p1, p0, rowname="Total")
, Test = perfminus(perfplus(p1, p0), p0, rowname="Test")
))
rbind(
  sperf(perfplus(p1, p0))
, sperf(p2)
)
}
\seealso{
\code{\link{perfplus}}
}
