% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfmon}
\alias{perfmon}
\title{Performance Monitor}
\usage{
perfmon(refsec = 5, monsec = 1, maxsec = Inf)
}
\arguments{
\item{refsec}{number of seconds to generate baseline estimate before monitoring}

\item{monsec}{number of seconds to sleep between reporting change}

\item{maxsec}{maximum number of seconds to monitor}
}
\value{
not returned but printed: a matrix reporting percentages with 2 rows
\item{curr/max}{Percent current values relative to maximum values: allows to monitor counter overflow (<<100 expected)}
\item{diff/base}{Percent incremental values relative to baseline: allows to monitor current energy (>=100 expected)}
and 4 columns
\item{package}{RAPL PACKAGE (socket including CORE+UNCORE+REST)}
\item{package}{RAPL CORE (CPU only)}
\item{package}{RAPL UNCORE (GPU only)}
\item{package}{RAPL DRAM (RAM only)}
}
\description{
A simple Performance Monitor
}
\details{
The first `refsecs` seconds it generates a baseline measurement,
and then reports each `monsecs` seconds until interrupted
}
\examples{
perfmon(refsec=30, monsec=1, maxsec=3)
}
\seealso{
\code{\link{perfnow}}, \code{\link{perfmax}}, \code{\link{perf}}
}
