% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfnames}
\alias{perfnames}
\title{Measurement names (without units)}
\usage{
perfnames(x)
}
\arguments{
\item{x}{a a vector of performance names such as \code{\link{colnames}} from \code{\link{perf}}}
}
\value{
a vector of names without units
}
\description{
Returns Measurement names without units
}
\examples{
p <- aperf(Insertionsort(runif(10000)))
data.frame(perf=colnames(p), labels=perflabels(colnames(p)), labels=perfnames(colnames(p)), labels=perfunits(colnames(p)))
p
colnames(p) <- perfnames(colnames(p))
p
}
\seealso{
\code{\link{perfnames}} and \code{\link{perflabels}}, \code{\link{perf}}
}
