% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfnow}
\alias{perfnow}
\title{Get performance counters}
\usage{
perfnow()
}
\value{
a double vector with four elements
\item{time}{\code{\link{proc.time}[[3]]} in seconds}
\item{package}{PACKAGE counter in Joule}
\item{core}{CORE counter in Joule}
\item{uncore}{UNCORE counter in Joule}
\item{dram}{DRAM counter in Joule}
}
\description{
Return the current powercap RAPL counters
}
\examples{
perfnow()
}
\seealso{
\code{\link{perfmax}}, \code{\link{perfmon}}, \code{\link{perf}}
}
