% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfsleep}
\alias{perfsleep}
\title{Measure background energy}
\usage{
perfsleep(x)
}
\arguments{
\item{x}{number of seconds to sleep}
}
\value{
like \code{\link{rawperf}}
}
\description{
Measure background energy of a sleep time
}
\examples{
perfsleep(1)

}
\seealso{
\code{\link{perfcalibrate}} and \code{\link{perf}}
}
