% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfthen}
\alias{perfthen}
\title{Measure end of interval and return perf}
\usage{
perfthen(x, rowname = NA, n = NA, b = 8, p = 1, t = 1, size = NA)
}
\arguments{
\item{x}{a measurement of interval begin returned by \code{\link{perfnow}}}

\item{rowname}{optional algoname for the returned matrix}

\item{n}{optional number of elements}

\item{b}{optional number of bytes per element}

\item{p}{optional number of processes}

\item{t}{optional number of threads}

\item{size}{optional \%RAM}
}
\value{
one-row matrix like \code{\link{rawperf}}
}
\description{
Return time in seconds and energies in Joule since interval begin `x`
}
\details{
Handles RAPL counter overflow assuming it won't happen twice
}
\examples{
message('Custom idle measurement')
p <- perfnow()
Sys.sleep(2)
perfthen(p)
}
\seealso{
\code{\link{perfcalibrate}} and \code{\link{perf}}

\code{\link{perfmax}}, \code{\link{perfmon}}, \code{\link{perf}}
}
