% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{perfunits}
\alias{perfunits}
\title{Measurement names with units}
\usage{
perfunits(x)
}
\arguments{
\item{x}{a a vector of performance names such as \code{\link{colnames}} from \code{\link{perf}}}
}
\value{
a vector of names with units
}
\description{
Returns long Measurement names including units
}
\examples{
p <- aperf(Insertionsort(runif(10000)))
data.frame(perf=colnames(p), labels=perflabels(colnames(p)), labels=perfnames(colnames(p)), labels=perfunits(colnames(p)))
p
colnames(p) <- perfunits(colnames(p))
t(p)
}
\seealso{
\code{\link{perfnames}} and \code{\link{perflabels}}, \code{\link{perf}}
}
