% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prunif.R
\name{prunif}
\alias{prunif}
\title{prunif}
\usage{
prunif(n, min = 0, max = .Machine$integer.max, pgrain = perfcores())
}
\arguments{
\item{n}{number of values to generate}

\item{min}{lower bound}

\item{max}{upper bound}

\item{pgrain}{parallelGrain paramter, default \code{\link{perfcores}}}
}
\value{
a random vector of doubles between min an max
}
\description{
Parallel random number generation
}
\details{
Using \code{\link[rTRNG]{runif_trng}}
}
\examples{
\dontrun{
require(RcppParallel)
require(rTRNG)
RcppParallel::setThreadOptions(numThreads = perfcores());
prunif(2^28)
}

}
\seealso{
\code{\link{runif}}, \code{\link[rTRNG]{runif_trng}}
}
