% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Quickpartright}
\alias{Quickpartright}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'quickpr'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Quickpartright(
  x,
  k = length(x)/2,
  situation = c("insitu", "exsitu"),
  method = c("index")
)
}
\arguments{
\item{x}{a double vector for searching (re-arranged with situation='insitu')}

\item{k}{the position within sorted x for which we ask the value}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
the value \code{x[k]}  as a double scalar with an attribute \code{\link{perf}} like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
y <- x[]; q <- Quickpartright(y)
y <- x[]; q <- Quickpartleft(y)
y <- x[]; q <- Zackpartright(y)
y <- x[]; q <- Quickselect(y)
q
}
\seealso{
\code{\link{quantile}} for R's standard quantiles,
\code{\link{Quickselect}} for select
\code{\link{Quickpart}} for partial sorting on both ends
\code{\link{Quickpartleft}} for partial sorting on the left end
\code{\link{Zackpartright}} for the greeNsort version giving more informative output
}
