% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Quicksort2}
\alias{Quicksort2}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Quick2'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Quicksort2(
  x,
  situation = c("insitu", "exsitu"),
  method = c("index", "pointer")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Picksort(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(Dupisort(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(RQuicksort2(y))}
,{y <- x[]; sperf(SQuicksort2(y))}
,{y <- x[]; sperf(Chicksort(y))}
,{y <- x[]; sperf(Zacksort(y))}
)
x <- testdb$func$tiesqrt(n)
rbind(
 {y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(Dupisort(y))}
,{y <- x[]; sperf(Chicksort(y))}
,{y <- x[]; sperf(Zacksort(y))}
)
}
\seealso{
\code{\link{Picksort}} for a version that uses  \code{\link{approxMedian}} pivots
\code{\link{Quicksort3}} for Bentley&McIlroy's algorithm that invests extra operations for isolating a (third) partition with pivot-ties for early termination.
\code{\link{Dupisort}} for Yaroslavskyi's 'Dual-Pivot-Quicksort' which invests extra operations for isolating a true third partition
\code{\link{Quicksort2B}} for a version that is tuned to avoid branch-(mis)predictions
\code{\link{Chicksort}} for a related algorithm that avoids tie swaps
\code{\link{Zocksort}} for an asymmetric algorithm that provides early termination on ties but is vulnerable to certain input patterns
\code{\link{Zacksort}} for an algorithm with similar costs for non-tied keys that provides early termination on ties
\code{\link{SQuicksort2}} for a stabilized version (directly)
\code{\link{RQuicksort2}} for a stabilized version (indirectly)
\code{\link{WQuicksort2}} for a version sorting strings indirectly (stablized)
\code{\link{PQuicksort2}} for a branch-parallel version
}
