% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Quicksort3}
\alias{Quicksort3}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Quick3'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Quicksort3(
  x,
  situation = c("insitu", "exsitu"),
  method = c("index", "pointer")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Dupisort(y))}
,{y <- x[]; sperf(Zacksort(y))}
)
x <- testdb$func$tiesqrt(n)
rbind(
 {y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Dupisort(y))}
,{y <- x[]; sperf(Zacksort(y))}
)
}
\seealso{
\code{\link{Quicksort2}} for Hoare's original algorithm without tuning for early termination on ties.
\code{\link{Dupisort}} for Yaroslavskyi's 'Dual-Pivot-Quicksort' which invests extra operations for isolating a true third partition
\code{\link{Zacksort}} for a algorithm with lower costs for non-tied keys that also provides early termination on ties
}
