% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{retperf}
\alias{retperf}
\title{Make perf return value}
\usage{
retperf(x, rowname = "")
}
\arguments{
\item{x}{the return value of a C-level sorting call with an attribute `perf`}

\item{rowname}{the rowname to be used}
}
\value{
measurements as a matrix with columns
\item{n}{number of elements}
\item{b}{(average) number of bytes per element}
\item{p}{number of cores}
\item{size}{average \%RAM}
\item{secs}{runTime in Seconds}
\item{base}{base energy in Joules}
\item{core}{core energy in Joules}
\item{unco}{uncore (GPU) energy in Joules}
\item{dram}{DRAM energy in Joules}
}
\description{
takes a return value with measurements of the R/C-glue layer and returns it with proper \code{\link{dim}} and \code{\link{dimnames}}
}
