% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntile2.R
\name{Rntile2_asc_asc_right_to_left_expensive}
\alias{Rntile2_asc_asc_right_to_left_expensive}
\title{ntiles of two sorted arrays from RIGHT to LEFT}
\usage{
Rntile2_asc_asc_right_to_left_expensive(L, R, k = (Ln + Rn)\%/\%2)
}
\arguments{
\item{L}{left sorted vector (ascending)}

\item{R}{right sorted vector  (ascending)}

\item{k}{number of elements to be separated on the RIGHT}
}
\value{
two-element vector c(l,r) giving the number of elements from L and R
}
\description{
expensive version based on full sorting
}
\examples{
Ln = 2^12
Rn = 2^12
N = Ln+Rn
skew = 2^7
L = rchisq(Ln, N-skew)
R = rchisq(Rn, N-skew)
L = L[order(current_keys(L))]
R = R[order(current_keys(R))]
k = N\%/\%4
Rntile2_asc_asc_right_to_left_expensive(L,R,k=k)

}
