% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntile2.R
\name{Rntile2_asc_revasc_right_to_left}
\alias{Rntile2_asc_revasc_right_to_left}
\title{ntiles of two sorted arrays (ASC REVASC from RIGHT to LEFT)}
\usage{
Rntile2_asc_revasc_right_to_left(L, R, k = (Ln + Rn)\%/\%2, verbose = FALSE)
}
\arguments{
\item{L}{left sorted vector (ascending)}

\item{R}{right sorted vector  (reverse ascending)}

\item{k}{number of elements to be separated on the LEFT}

\item{verbose}{TRUE for reporting the steps}
}
\value{
two-element vector c(l,r) giving the number of elements from L and R
}
\description{
ntiles of two sorted arrays (ASC REVASC from RIGHT to LEFT)
}
\examples{
ycol  = c("red","blue")
Ln = 2^12
Rn = 2^12
N = Ln+Rn
skew = runif(1, -2^8, 2^8)
L = rchisq(Ln, N-skew)
R = rchisq(Rn, N+skew)
L = L[order(current_keys(L))]
R = R[order(current_keys(R))]
ylim = range(c(L,R))
oldpar <- par(mfrow=c(2,2))
k = as.integer(N*1/12)
xlr = Rntile2_asc_asc_right_to_left_expensive(L, R, k=k)
lr = Rntile2_asc_revasc_right_to_left(L,rev(R),k=k)
stopifnot(identical(lr, xlr))
plot(L, ylim=ylim, col=ycol[1+(seq_along(L)<(Ln-lr[1]+1))])
m = min(rev(L)[lr[1]], rev(R)[lr[2]])
abline(h=m, col=ycol[1])
plot(rev(R), ylim=ylim, col=ycol[1+(rev(seq_along(R))<(Rn-lr[2]+1))])
abline(h=m, col=ycol[1])
k = as.integer(N*11/12)
xlr = Rntile2_asc_asc_right_to_left_expensive(L, R, k=k)
lr = Rntile2_asc_revasc_right_to_left(L,rev(R),k=k)
stopifnot(identical(lr, xlr))
plot(L, ylim=ylim, col=ycol[1+(seq_along(L)<(Ln-lr[1]+1))])
m = min(rev(L)[lr[1]], rev(R)[lr[2]])
abline(h=m, col=ycol[1])
plot(rev(R), ylim=ylim, col=ycol[1+(rev(seq_along(R))<(Rn-lr[2]+1))])
abline(h=m, col=ycol[1])
par(oldpar)

}
