% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{RQuicksort2}
\alias{RQuicksort2}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'RQuick2'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
RQuicksort2(
  x,
  situation = c("insitu", "exsitu"),
  reorder_inplace = FALSE,
  method = c("index")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{reorder_inplace}{\code{FALSE} allocates 100\% buffer to speed up final reordering, \code{TRUE} reorders inplace}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(RQuicksort2(y))}
,{y <- x[]; sperf(SQuicksort2(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(RQuicksort2B(y))}
)
}
\seealso{
\code{\link{SQuicksort2}} for a stable version that avoids random access
\code{\link{Quicksort2}} for the unstable version
\code{\link{RQuicksort2B}} for a version that is tuned to avoid branch-(mis)prediction
}
