% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{secs}
\alias{secs}
\alias{runTime}
\title{Extract runTime}
\usage{
secs(x)

runTime(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'runTime'
}
\description{
Returns the runTime in seconds
}
\details{
runTime is elapsed time, it grows if concurring processes delay processing.
}
\section{Functions}{
\itemize{
\item \code{runTime()}: alternative function name for `secs`

}}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
secs(ret)
runTime(ret)
}
\seealso{
\code{\link{perf}}
}
