% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{size}
\alias{size}
\alias{\%RAM}
\title{Extract \%RAM}
\usage{
size(x)

`\%RAM`(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} '%RAM'
}
\description{
Returns the average RAM consumption over \code{\link{runTime}} divided by data size
}
\section{Functions}{
\itemize{
\item \code{`\%RAM`()}: alternative function name for `size`

}}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
size(ret)
"\%RAM"(ret)
}
\seealso{
\code{\link{perf}}
}
