% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{sizesecs}
\alias{sizesecs}
\alias{tFootprint}
\title{Extract tFootprint}
\usage{
sizesecs(x)

tFootprint(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'tFootprint'
}
\description{
Returns the tFootprint in %RAM times seconds
}
\details{
This is \code{\link{size}} times \code{\link{secs}}
}
\section{Functions}{
\itemize{
\item \code{tFootprint()}: alternative function name for `sizesecs`

}}
\examples{
ret <- Insertionsort(runif(1000))
perf(ret)
sizesecs(ret)
tFootprint(ret)
}
\seealso{
\code{\link{perf}}
}
