% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{SQuicksort2B}
\alias{SQuicksort2B}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'SQuick2B'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
SQuicksort2B(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(SQuicksort2B(y))}
,{y <- x[]; sperf(RQuicksort2B(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(SQuicksort2(y))}
,{y <- x[]; sperf(KatasortB(y))}
,{y <- x[]; sperf(Frogsort1B(y))}
)
}
\seealso{
\code{\link{RQuicksort2B}} for an algorithm using pointers but random access
\code{\link{Quicksort2B}} for the unstable algorithm
\code{\link{SQuicksort2}} for a version that is not tuned to avoid branch-(mis)prediction
\code{\link{KatasortB}} for a stable tuned mergesort using 100\% buffer
\code{\link{Frogsort1B}} for a stable tuned algorithm using 50\% buffer
}
