% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Squidsort2}
\alias{Squidsort2}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Squid2'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Squidsort2(
  x,
  situation = c("insitu", "exsitu"),
  method = c("pointer"),
  p = 1/7
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}

\item{p}{fraction of  buffer relative to data size}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Frogsort2(y))}
,{y <- x[]; sperf(Frogsort2A(y))}
,{y <- x[]; sperf(Squidsort1(y))}
,{y <- x[]; sperf(Squidsort2(y))}
,{y <- x[]; sperf(Timsort(y))}
,{y <- x[]; sperf(gfxTimsort(y))}
)
x <- testdb$func$descall(n)
rbind(
 {y <- x[]; sperf(Frogsort2(y))}
,{y <- x[]; sperf(Frogsort2A(y))}
,{y <- x[]; sperf(Squidsort1(y))}
,{y <- x[]; sperf(Squidsort2(y))}
,{y <- x[]; sperf(Timsort(y))}
,{y <- x[]; sperf(gfxTimsort(y))}
)
}
\seealso{
\code{\link{Frogsort2}} for the undelying untuned version
\code{\link{Frogsort2A}} for the undelying tuned version
\code{\link{Squidsort1}} for an balanced version using 50\% buffer
\code{\link{Timsort}} for an algorithm tuned for the best cases of presorting
\code{\link{gfxTimsort}} for an algorithm tuned for the best cases of presorting
}
