% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{stable_keys}
\alias{stable_keys}
\title{stable_keys}
\usage{
stable_keys(x)
}
\arguments{
\item{x}{a double or character vector to be rounded (or not)}
}
\value{
The rounded
}
\description{
Convert Keys to Keys used in stable testing
}
\details{
doubles are truncated towards zero
strings must be shorter equal 20 and are truncated from the first dot to the right
}
\examples{
  stable_keys(seq(-10, 10, 0.5))
  stable_keys(num2str(seq(-10, 10, 0.5)))
}
\seealso{
\code{\link{greeNsort-package}}
}
