% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{StorksortB}
\alias{StorksortB}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'StorkB'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
StorksortB(x, situation = c("exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Swansort(y, situation = "exsitu"))}
,{y <- x[]; sperf(SwansortB(y, situation = "exsitu"))}
,{y <- x[]; sperf(Storksort(y, situation = "exsitu"))}
,{y <- x[]; sperf(StorksortB(y, situation = "exsitu"))}
,{y <- x[]; sperf(Frogsort1(y, situation = "exsitu"))}
,{y <- x[]; sperf(Frogsort1B(y, situation = "exsitu"))}
,{y <- x[]; sperf(Frogsort2(y, situation = "exsitu"))}
,{y <- x[]; sperf(Frogsort2B(y, situation = "exsitu"))}
)
}
\seealso{
\code{\link{Storksort}} for the untuned version,
\code{\link{SwansortB}} for the similar tuned more general t-partition P&P algorithm which needs 100\% buffer,
\code{\link{Frogsort1B}} for the similar tuned much more general S&M algorithm which also needs 50\% buffer.
\code{\link{Frogsort2}} for the fastest S&M algorithm which also needs even less buffer and is even faster.
}
