% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{string_check}
\alias{string_check}
\title{string check}
\usage{
string_check(x)
}
\arguments{
\item{x}{a vector of \code{\link{character}}}
}
\value{
vector of pos < nchar(x), where pos==0 for strings that are OK
}
\description{
find positions in string that equal the BUFFER_CHAR
}
\examples{
stopifnot(all(string_check(vtestdb$func$KafkaWords())==0))
stopifnot(all(string_check(vtestdb$func$KafkaParts())==0))
stopifnot(all(string_check(vtestdb$func$BibleWords())==0))
stopifnot(all(string_check(vtestdb$func$BibleVerses())==0))
stopifnot(all(string_check(vtestdb$func$DictWords())==0))
}
