% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Swansort}
\alias{Swansort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Swan'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Swansort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(DietCPTsort(y, situation = "insitu"))}
,{y <- x[]; sperf(Kiwisort(y, situation = "insitu"))}
,{y <- x[]; sperf(KiwisortB(y, situation = "insitu"))}
,{y <- x[]; sperf(Swansort(y, situation = "insitu"))}
,{y <- x[]; sperf(SwansortB(y, situation = "insitu"))}
,{y <- x[]; sperf(Storksort(y, situation = "'insitu'"))}
)
rbind(
 {y <- x[]; sperf(DietCPTsort(y, situation = "exsitu"))}
,{y <- x[]; sperf(Kiwisort(y, situation = "exsitu"))}
,{y <- x[]; sperf(KiwisortB(y, situation = "exsitu"))}
,{y <- x[]; sperf(Swansort(y, situation = "exsitu"))}
,{y <- x[]; sperf(SwansortB(y, situation = "exsitu"))}
,{y <- x[]; sperf(Storksort(y, situation = "exsitu"))}
,{y <- x[]; sperf(StorksortB(y, situation = "exsitu"))}
)
}
\seealso{
\code{\link{DietCPTsort}} for a t-partition algorithm with an extra counting pass,
\code{\link{Kiwisort}} for a ping-pong algorithm (faster in 'insitu' setting)
\code{\link{Storksort}} for a t-partition algorithm which requires only 50\% buffer,
\code{\link{SwansortB}} for a version tuned to avoid branch-(mis)prediction.
}
