% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\docType{data}
\name{testdb}
\alias{testdb}
\title{Testdata database (doubles)}
\format{
A data frame with one row per test generator and these columns:
\describe{
  \item{ name }{ full test name }
  \item{ label }{ abbreviated test label}
  \item{ desc }{ a description of the test }
  \item{ major }{ one letter classifiying the test as 'P'=Permutation, 'A' = Ascending, 'D' = Descending' 'a' = AscendingDescending, 'd'= DescendingAscending, 'T'= Ties, 'B' = Ascending with ties, 'S' = Stability }
  \item{ minor }{ one letter identifying the test }
  \item{ stable }{ \code{TRUE} if pattern is suitable for stability testing with \code{\link{current_keys}} }
  \item{ entropy }{ entropy }
  \item{ relentropy }{ relative entropy }
  \item{ monotAscFrac }{ fraction of ascending differences }
  \item{ strictAscFrac }{  fraction of strictly ascending differences }
  \item{ spearman }{ Spearman correation between position and value }
  \item{ code }{ the test generator code }
  \item{ func }{ the callable test generator function }
  }
}
\usage{
testdb
}
\description{
A dataset describing all testdata generators in this package.
}
\examples{
  table(testdb$major)
  summary(testdb$entropy)
  summary(testdb$relentropy)
  summary(testdb$monotAscFrac)
  summary(testdb$strictAscFrac)
  summary(testdb$spearman)
  \donttest{
    View(testdb)
  }

}
\seealso{
\code{\link{vtestdb}} and \code{\link{algodb}}
}
\keyword{datasets}
