% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testpart}
\alias{testpart}
\title{Test partial sorting algorithms}
\usage{
testpart(
  R = 0,
  N = c(1:7, insertionsort_limit() - 1L, insertionsort_limit(), insertionsort_limit() +
    1L, 127L, 128L, 129L, 255L, 256L, 257L),
  S = c("insitu", "exsitu"),
  A = algodb[algodb$task \%in\% c("partial", "PARTIAL"), "label"],
  D = c("ascall", "descall", "permut", "tiesqrt", "tielog2"),
  out = file(paste("test", r, sep = ""), open = "w"),
  interrupt = out == "",
  verbose = !interrupt
)
}
\arguments{
\item{R}{vector of random seeds for replications}

\item{N}{vector of samplesizes}

\item{S}{vector of situations in c("insitu","exsitu"), non-existing will be skipped automatically}

\item{A}{vector of algorithm labels from \code{\link{algodb}}}

\item{D}{vector of data labels from \code{\link{testdb}}}

\item{out}{output destination: "" for console or filename}

\item{interrupt}{TRUE for interrupting at first error, FALSE for continuing all tests}

\item{verbose}{TRUE to cat all tests, FALSE for silent testing}
}
\value{
TRUE if all tests were OK otherwise FALSE
}
\description{
We assume that partial sorting is unstable (but handle the case of stable testing where `current_keys(0.5) != 0.5` and the test data have digits after the decimal dot)
}
\examples{
testpart(out = "", verbose = TRUE)
}
\seealso{
\code{\link{current_keys}}, \code{\link{testselect}}, \code{\link{testsort}}
}
