% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{testsort}
\alias{testsort}
\title{Test sorting algorithms}
\usage{
testsort(
  R = 0,
  N = c(1:7, insertionsort_limit() - 1L, insertionsort_limit(), insertionsort_limit() +
    1L, 127L, 128L, 129L, 255L, 256L, 257L),
  M = c("pointer", "index"),
  S = c("insitu", "exsitu"),
  A = algodb[algodb$task \%in\% c("sort", "SORT") & !(algodb$label \%in\% c("IPS4o",
    "IS4o")), "label"],
  D = testdb$label,
  T = 1:perfcores(),
  out = file(paste("test", r, sep = ""), open = "w"),
  interrupt = out == "",
  verbose = !interrupt
)
}
\arguments{
\item{R}{vector of random seeds for replications}

\item{N}{vector of samplesizes}

\item{M}{vector of methods in c("index","pointer"), non-existing will be skipped automatically}

\item{S}{vector of situations in c("insitu","exsitu"), non-existing will be skipped automatically}

\item{A}{vector of algorithm labels from \code{\link{algodb}}}

\item{D}{vector of data labels from \code{\link{testdb}}}

\item{T}{scalar with number of parallel cores}

\item{out}{output destination: "" for console or filename}

\item{interrupt}{TRUE for interrupting at first error, FALSE for continuing all tests}

\item{verbose}{TRUE to cat all tests, FALSE for silent testing}
}
\value{
TRUE if all tests were OK otherwise FALSE
}
\description{
Test sorting algorithms
}
\examples{
testsort(out = "", verbose = TRUE)
}
\seealso{
\code{\link{current_keys}}, \code{\link{testselect}}, \code{\link{testpart}}, \code{\link{vtestsort}}
}
