% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Tricksort}
\alias{Tricksort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Trick'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Tricksort(x, situation = c("insitu", "exsitu"), method = c("index", "pointer"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Tricksort(y))}
,{y <- x[]; sperf(Dupisort(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(ZacksortB(y))}
)
x <- testdb$func$tiesqrt(n)
rbind(
 {y <- x[]; sperf(Tricksort(y))}
,{y <- x[]; sperf(Dupisort(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(ZacksortB(y))}#'
)
}
\seealso{
\code{\link{Dupisort}} for Yaroslavskyi's version (but with random pivots and still with redundant operations)
\code{\link{Quicksort2}} for Hoare's binary algorithm without tuning for early termination on ties.
\code{\link{Quicksort3}} for Bentley&McIlroy's slower threeway algorithm with tuning for early termination on ties.
\code{\link{Zacksort}} for a simpler binary algorithm with low costs for non-tied keys that also provides early termination on ties but is easier parallelized
\code{\link{Quicksort2B}} and \code{\link{ZacksortB}} as a benchmark
}
