% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{UInsertionsort}
\alias{UInsertionsort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'UIns'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
UInsertionsort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- vtestdb$func$KafkaWords()[1:n]
rbind(
 {y <- x[]; sperf(UInsertionsort(y))}
,{y <- x[]; sperf(WInsertionsort(y))}
)
}
\seealso{
\code{\link{Insertionsort}} for the standard,
\code{\link{UZacksort}} for an algorithm that uses \code{UInsertionsort}
\code{\link{WInsertionsort}} for a version that is explicitely stabilized
\code{\link{VInsertionsort}} for a version sorts elements of varying size directly (moving elements, not pointers)
\code{\link{RInsertionsort}} for a version that sorts doubles via pointers
}
