% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perf.R
\name{unco}
\alias{unco}
\title{Extract GPU Energy}
\usage{
unco(x)
}
\arguments{
\item{x}{see \code{\link{perf}}}
}
\value{
a double column with \code{\link{colnames}} 'unco'
}
\description{
Returns the RAPL uncore Energy in Joule
}
\details{
\code{\link{unco}} is the RAPL uncore energy
}
\examples{
ret <- Insertionsort(runif(1000))
ret
core(ret)
unco(ret)
}
\seealso{
\code{\link{retperf}}, \code{\link{base}}, \code{\link{core}}, \code{\link{unco}}, \code{\link{dram}}
}
