% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{UZacksortB}
\alias{UZacksortB}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'UZackB'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
UZacksortB(x, situation = c("exsitu", "insitu"), method = c("index"))
}
\arguments{
\item{x}{a character vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
x <- vtestdb$func$KafkaWords()
rbind(
 {y <- x[]; sperf(UZacksort(y))}
,{y <- x[]; sperf(WQuicksort2(y))}
,{y <- x[]; sperf(UZacksortB(y))}
,{y <- x[]; sperf(WQuicksort2B(y))}
,{y <- x[]; sperf(UKnuthsort(y))}
,{y <- x[]; sperf(VKnuthsort(y))}
,{y <- x[]; sperf(VFrogsort1(y))}
,{y <- x[]; sperf(VKnuthsortA(y))}
,{y <- x[]; sperf(VFrogsort1A(y))}
)
}
\seealso{
\code{\link{WQuicksort2B}} for a tuned indirect stabilized quicksort using pointers  to size-varying elements (strings, no ties by definition),
\code{\link{UZacksort}} for the untuned version,
\code{\link{RQuicksort2B}} for a tuned indirect Quicksort for equally-sized elements,
\code{\link{ZacksortB}} for the version for equally-sized elements,
}
