% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{VKnuthsort}
\alias{VKnuthsort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'VKnuth'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
VKnuthsort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a character vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
x <- vtestdb$func$KafkaWords()
rbind(
 {y <- x[]; sperf(UZacksort(y))}
,{y <- x[]; sperf(WQuicksort2(y))}
,{y <- x[]; sperf(UZacksortB(y))}
,{y <- x[]; sperf(WQuicksort2B(y))}
,{y <- x[]; sperf(UKnuthsort(y))}
,{y <- x[]; sperf(VKnuthsort(y))}
,{y <- x[]; sperf(VFrogsort1(y))}
,{y <- x[]; sperf(VKnuthsortA(y))}
,{y <- x[]; sperf(VFrogsort1A(y))}
)
}
\seealso{
\code{\link{UZacksort}} for indirect Zacksort using pointers to size-varying elements (strings, not stable),
\code{\link{WQuicksort2}} for indirect stabilized quicksort using pointers  to size-varying elements (strings, no ties by definition),
\code{\link{UKnuthsort}} for an indirect mergesort of size-varying elements (strings) using Knuth's merge with one loop check,
\code{\link{VFrogsort1}} for direct Frogsort of size-varying elements (strings),
\code{\link{VKnuthsortA}} for an adaptive tuned version,
\code{\link{Knuthsort}} for the version for equally-sized elements,
}
