% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{vtestsort}
\alias{vtestsort}
\title{Varying Test sorting algorithms}
\usage{
vtestsort(
  R = 0,
  M = c("pointer", "index"),
  S = c("insitu", "exsitu"),
  A = algodb[algodb$varying & algodb$task \%in\% c("sort", "SORT"), "label"],
  D = vtestdb$label,
  out = file(paste("test", r, sep = ""), open = "w"),
  interrupt = out == "",
  verbose = !interrupt
)
}
\arguments{
\item{R}{vector of random seeds for replications}

\item{M}{vector of methods in c("index","pointer"), non-existing will be skipped automatically}

\item{S}{vector of situations in c("insitu","exsitu"), non-existing will be skipped automatically}

\item{A}{vector of algorithm labels from \code{\link{algodb}}}

\item{D}{vector of data labels from \code{\link{vtestdb}}}

\item{out}{output destination: "" for console or filename}

\item{interrupt}{TRUE for interrupting at first error, FALSE for continuing all tests}

\item{verbose}{TRUE to cat all tests, FALSE for silent testing}
}
\value{
TRUE if all tests were OK otherwise FALSE
}
\description{
Varying Test sorting algorithms
}
\examples{
vtestsort(out = "", verbose = TRUE)
}
\seealso{
\code{\link{current_keys}}, \code{\link{testsort}}
}
