% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Zackpart}
\alias{Zackpart}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'zackpt'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Zackpart(
  x,
  partial = c(1L, length(x)),
  situation = c("insitu", "exsitu"),
  method = c("index")
)
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{partial}{a double vector with two elements for lower and upper limit of partial sorting}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$tiesqrt(n)
y <- x[]; sperf(Quickpart(y, c(0.33, 0.67)*n)); plot(y)
y <- x[]; sperf(Zackpart(y, c(0.33, 0.67)*n)); plot(y)
y <- x[]; sperf(Zuckpart(y, c(0.33, 0.67)*n)); plot(y)
}
\seealso{
\code{\link{Quickpart}} for the classic quick algorithm that does not guarantee to include all ties
\code{\link{Zuckpart}} for a version that switches direction only on the required branch
}
