% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{ZacksortB}
\alias{ZacksortB}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'ZackB'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
ZacksortB(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(ZacksortB(y))}
,{y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(Tricksort(y))}
,{y <- x[]; sperf(Frogsort1B(y))}
,{y <- x[]; sperf(Frogsort2(y))}
)
x <- testdb$func$tiesqrt(n)
rbind(
 {y <- x[]; sperf(ZacksortB(y))}
,{y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(Tricksort(y))}
,{y <- x[]; sperf(Frogsort1B(y))}
,{y <- x[]; sperf(Frogsort2(y))}
)
x <- testdb$func$ascnorm(n)
rbind(
 {y <- x[]; sperf(ZacksortB(y))}
,{y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2B(y))}
,{y <- x[]; sperf(Tricksort(y))}
,{y <- x[]; sperf(Frogsort1B(y))}
,{y <- x[]; sperf(Frogsort2(y))}
)
}
\seealso{
\code{\link{Zacksort}} for the untuned version
\code{\link{Quicksort2B}} for Hoare's quicksort tuned to avoid branch-(mis)prediction
\code{\link{Frogsort1B}} for a stable greeNsort algorithm
\code{\link{Tricksort}} for a slightly improved version avoiding some redundand operations in Yaroslavskyi's 'Dual-Pivot-Quicksort' which invests extra operations for isolating a true third partition
\code{\link{Frogsort2}} for the fastest stable greeNsort algorithm with O(N*log(N)) worst-case
\code{\link{UZacksortB}} for a version sorting strings indirectly (not stabilized, stabilization removes ties)
}
