% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Zickselect}
\alias{Zickselect}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'zick'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Zickselect(
  x,
  k = length(x)/2,
  situation = c("insitu", "exsitu"),
  method = c("index")
)
}
\arguments{
\item{x}{a double vector for searching (re-arranged with situation='insitu')}

\item{k}{the position within sorted x for which we ask the value}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
a integer pair of positions k_min, k_max with an attribute \code{\link{perf}} like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$tiesqrt(n)
k <- n\%/\%2
y <- x[]; p <- Pickselect(y, k)
z <- x[]; v <- Zickselect(z, k)
plot(sort(x), main="fully sorted data")
abline(h=p, col="blue")
abline(v=v, col=greensort_color)
abline(v=k, col="red")
plot(y, main="partially sorted data (Pickselect)")
abline(h=p, col="blue")
abline(v=v, col=greensort_color)
abline(v=k, col="red")
plot(z, main="partially sorted data (Zickselect)")
abline(h=p, col="blue")
abline(v=v, col=greensort_color)
abline(v=k, col="red")
}
\seealso{
\code{\link{quantile}} for R's standard quantiles,
\code{\link{Pickselect}} for a prior-art version giving less informative output
\code{\link{Zackselect}} for a usually faster version with random pivots
}
