% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{ZuckselectB}
\alias{ZuckselectB}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'zuckB'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
ZuckselectB(
  x,
  k = length(x)/2,
  situation = c("insitu", "exsitu"),
  method = c("index")
)
}
\arguments{
\item{x}{a double vector for searching (re-arranged with situation='insitu')}

\item{k}{the position within sorted x for which we ask the value}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
a integer pair of positions k_min, k_max with an attribute \code{\link{perf}} like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
message("use bigger n, eg.g. 2^20, to see timing difference")
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
y <- x[]; z <- Zuckselect(y)
y <- x[]; b <- ZuckselectB(y)
z
b
}
\seealso{
\code{\link{quantile}} for R's standard quantiles,
\code{\link{Zuckselect}} for the version without block-tuning
\code{\link{DuckselectB}} for a version that is adaptive to presorted data
}
