% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greeNsort.R
\name{Zucksort}
\alias{Zucksort}
\title{\Sexpr[echo=FALSE, results=rd, stage=build]{{thisalgo <- 'Zuck'; greeNsort::algodb[thisalgo,'name']}}}
\usage{
Zucksort(x, situation = c("insitu", "exsitu"), method = c("index"))
}
\arguments{
\item{x}{a \code{\link{double}} vector to be sorted}

\item{situation}{"insitu" will only allocate buffer memory and use the existing RAM for sorting, "exsitu" will allocate completely fresh RAM for data and buffer}

\item{method}{an attempt to classify the implementation,  "index" means the implementation rather indexes into arrays, "pointer" rather uses pointer arithmetic}
}
\value{
like \code{\link{rawperf}}
}
\description{
\Sexpr[echo=FALSE, results=rd, stage=build]{gsub("(.*)(like )(.+)( but)(.*)","\\\\1like \\\\\\\\code{\\\\\\\\link{\\\\3} } but\\\\5",greeNsort::algodb[thisalgo,'desc'])}
}
\details{
\Sexpr[echo=FALSE, results=text, stage=build]{greeNsort::algodb[thisalgo,'expl']}

\Sexpr[echo=FALSE, results=verbatim, stage=build]{{i <- !colnames(greeNsort::algodb) \%in\% c('func','desc','expl'); cat(paste(colnames(greeNsort::algodb)[i], ' - ',  greeNsort::algodb[thisalgo,i], '\n', sep=''), sep='')}}

See \code{\link{algodb}} for the complete table of algorithms.
}
\examples{
n <- getOption("greensort_example_size")
x <- testdb$func$permut(n)
rbind(
 {y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(ZucksortA(y))}
,{y <- x[]; sperf(Ducksort(y))}
,{y <- x[]; sperf(ZucksortB(y))}
,{y <- x[]; sperf(ZucksortD(y))}
)
x <- testdb$func$tiesqrt(n)
rbind(
 {y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(ZucksortA(y))}
,{y <- x[]; sperf(Ducksort(y))}
,{y <- x[]; sperf(ZucksortB(y))}
,{y <- x[]; sperf(ZucksortD(y))}
)
x <- testdb$func$ascnorm(n)
rbind(
 {y <- x[]; sperf(Zacksort(y))}
,{y <- x[]; sperf(Quicksort2(y))}
,{y <- x[]; sperf(Quicksort3(y))}
,{y <- x[]; sperf(ZucksortA(y))}
,{y <- x[]; sperf(Ducksort(y))}
,{y <- x[]; sperf(ZucksortB(y))}
,{y <- x[]; sperf(ZucksortD(y))}
)
}
\seealso{
\code{\link{Zacksort}} for a variant that switches partition directon on both branches
\code{\link{ZucksortA}} for a variant tuned for presorting before recursion
\code{\link{Ducksort}} for a variant tuned for presorting during recursion
\code{\link{ZucksortB}} for a block-tuned variant
\code{\link{ZucksortD}} for a variant using a deterministic pivot
}
